# docs/swagger.yaml
basePath: /
definitions:
...
  participants.InfoParticipant:
    properties:
      accepted:
        type: boolean
      charge:
        type: integer
    type: object
  participants.requestRegisterParticipant:
    properties:
      amount:
        type: number
    type: object
  participants.resParticipant:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/participants.InfoParticipant'
      error:
        type: boolean
      msg:
        type: string
      type:
        type: integer
    type: object
  participants.responseRegisterParticipant:
    properties:
      code:
        type: integer
      data:
        type: string
      error:
        type: boolean
      msg:
        type: string
      type:
        type: integer
    type: object
...
  validator.resRegisterVote:
    properties:
      code:
        type: integer
      data:
        type: string
      error:
        type: boolean
      msg:
        type: string
      type:
        type: integer
    type: object
  validator.rqRegisterVote:
    properties:
      hash:
        type: string
      wallet_id:
        type: string
    type: object
host: http://172.174.77.149:2054
info:
  contact:
    email: info@bjungle.net
    name: API Support
  description: Documentación del API que conecta con el core de BLion
  license:
    name: Software Owner
    url: https://www.bjungle.net/terms/licenses
  termsOfService: https://www.bjungle.net/terms/
  title: BLion
  version: "1.1"
paths:
...
  /api/v1/participants/info/{wallet}:
    get:
      consumes:
      - application/json
      description: Método para obtener la información de un participante por su wallet
        id
      parameters:
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Id de la wallet
        in: path
        name: wallet
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/participants.resParticipant'
      summary: Método para obtener la información de un participante
      tags:
      - Participants
  /api/v1/participants/register:
    post:
      consumes:
      - application/json
      description: Método para registrar un participante en la lotería actual
      parameters:
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Datos para registrarse en la lotería
        in: body
        name: requestRegisterParticipant
        required: true
        schema:
          $ref: '#/definitions/participants.requestRegisterParticipant'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/participants.responseRegisterParticipant'
      summary: Método para registrar un participante en la lotería actual
      tags:
      - Participants
...
  /api/v1/validators/register:
    post:
      consumes:
      - application/json
      description: Método para registrar el voto de un validador
      parameters:
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Datos para registrar el voto
        in: body
        name: rqRegisterVote
        required: true
        schema:
          $ref: '#/definitions/validator.rqRegisterVote'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/validator.resRegisterVote'
      summary: Método para registrar el voto de un validador
      tags:
      - Validator
swagger: "2.0"
...